% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmiHier_Lasso.R
\name{omihier_lasso}
\alias{omihier_lasso}
\title{OmiHier Lasso Function}
\usage{
omihier_lasso(data)
}
\arguments{
\item{data}{Limit its first column category variable (numeric type), 2-n column feature variable.}
}
\value{
This function does not return any value. It plots the OmiHier_Lasso dendrogram.
}
\description{
This function performs OmiHier analysis using lasso logistic regression.
}
\examples{
data_file <- system.file("data", "Faecal.csv", package = "OmiHier")
data <- read.csv(data_file)
omihier_lasso(data)

}
