% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmiHier_LightGBM.R
\name{omihier_lightgbm}
\alias{omihier_lightgbm}
\title{OmiHier LightGBM}
\usage{
omihier_lightgbm(data)
}
\arguments{
\item{data}{Limit its first column category variable (numeric type), 2-n column feature variable.}
}
\value{
This function does not return any value. It plots the OmiHier_LightGBM dendrogram.
}
\description{
This function performs OmiHier analysis using LightGBM for multiclass classification.
}
\examples{
data_file <- system.file("data", "Faecal.csv", package = "OmiHier",mustWork = TRUE)
data <- read.csv(data_file)
omihier_lightgbm(data)

}
