% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmiHier_SVM.R
\name{omihier_svm}
\alias{omihier_svm}
\title{OmiHier SVM}
\usage{
omihier_svm(data)
}
\arguments{
\item{data}{Limit its first column category variable (numeric type, limit it must start with 1), 2-n column feature variable.}
}
\value{
This function does not return any value. It plots the OmiHier_SVM dendrogram.
}
\description{
This function performs OmiHier analysis using SVM for multiclass classification.
}
\examples{
data_file <- system.file("data", "Faecal_svm.csv", package = "OmiHier",mustWork = TRUE)
data <- read.csv(data_file)
omihier_svm(data)

}
